notificationsAPI/requirements.txt
Flask
Flask-Migrate
gunicorn
httpx[http2]
pyodbc
pyjwt
cryptography
psycopg2
python-dotenv
requests